blocking

+++ database/migrations/2018_06_23_173314_create_menseuelle_commissions_table.php
@@ -19,7 +19,7 @@ class CreateMenseuelleCommissionsTable extends Migration
             // Schema declaration
             $table->string('type');
             $table->integer('id_account');
-            $table->float('amount');
+            $table->double('amount');
             $table->timestamps();
             // Constraints declaration
         });
